moved createUpdateIndexFile function to index_creation.js
