+++ index_creation.js
@@ -1,13 +1,73 @@
 import fs from 'fs/promises';
 import path from 'node:path';
 import * as cheerio from 'cheerio';
+import { confirm } from '@inquirer/prompts';
+import chalk from 'chalk';
+import { getConfig } from './config.js';
 import logger from './logger.js';
-import { fileExists, truncateFile } from './helpers.js';
+import { fileExists, getHtmlFiles, truncateFile } from './helpers.js';
 
 /**
  * Namespace for functions used to extract text from HTML files and create a file index
  * @namespace fileIndex
  */
+
+/**
+ * Create or update an index file with all source files contained in the original data source
+ * @async
+ * @function createUpdateIndexFile
+ * @memberof fileIndex
+ * @requires confirm
+ * @requires helpers
+ * @requires index_creation
+ * @returns {Promise<Object>} Object containing the list of files as an array, and the index file path
+ */
+export async function createUpdateIndexFile() {
+  try {
+    const config = getConfig();
+    // Prompt user to confirm the directory for the source files of the data source
+    config.OK = await confirm({
+      message: `The directory for the files of data source '${config.dataSourceSelection}' is '${config.FILE_DIR}'. Continue?`,
+    });
+
+    // Throw error if user cancels the operation due to incorrect configuration
+    if (!config.OK) {
+      logger.info(`User cancelled the operation due to configuration errors`);
+      console.log(
+        `${chalk.red('X')}` +
+          ` Operation cancelled. Check configuration in ${chalk.cyan('./config.yaml')}.`
+      );
+      throw new Error('Operation cancelled');
+    }
+
+    const fileList = await getHtmlFiles(config.FILE_DIR);
+    console.log(
+      `  Found ${chalk.cyan(fileList.length)} HTML files in data source directory. Creating index file...`
+    );
+    logger.info(
+      `Found ${fileList.length} HTML files in data source directory. Creating index file.`
+    );
+    const indexFile = await createFileIndex({
+      directoryPath: config.FILE_DIR,
+      fileList,
+    });
+    console.log(
+      `${chalk.green('√')}` +
+        ` Index file for '${config.dataSourceSelection}' source data created at ${chalk.cyan(indexFile)}`
+    );
+    logger.info(
+      `Index file for '${config.dataSourceSelection}' source data created at ${indexFile}`
+    );
+    return { fileList, indexFile };
+  } catch (error) {
+    if (error.message === 'Operation cancelled') {
+      return { fileList: [], indexFile: '' };
+    } else {
+      logger.error(`createUpdateIndexFile() Error:\n${error}`);
+      throw new Error(`Failed to create index file: ${error.message}`);
+    }
+  }
+}
 
 /** Creates a JSON index of all files in a specified directory and saves it to a file
  * <br> The index includes the filename, title, and breadcrumbs (optional) extracted from the HTML contents
